# swagger.yaml
swagger: "2.0"
info:
  description: "Description..."
  version: "Alpha 1"
  title: "Easy Lightning REST API"
  # termsOfService: "http://swagger.io/terms/"
  contact:
    email: "contact@mail.com"
  license:
    name: "GNU General Public License v3.0"
    url: "https://www.gnu.org/licenses/gpl-3.0.html"
host: "localhost:8118"
basePath: "/"
...
schemes:
- "http"
paths:
  /invoice/createInvoice:
    get:
      tags:
      - "/invoice"
      summary: "Create an invoice to pay it"
      operationId: "createInvoice"
      consumes: 
      - "application/json"
      produces: 
      - "application/json"
      parameters:
      - in: "query"
        name: "amount"
        required: true
        type: "integer"
        default: 0
        description: "Amount of Satoshi"
      - in: "query"
        name: "memo"
        required: false
        type: "string"
        description: "Add a memo to the invoice"
      - in: "query"
        name: "expiry"
        required: false
        type: "integer"
        description: "Expiry time in seconds. Default: 3600 (1 Hour)"
      responses:
        200:
          description: "Invoice successfully created"
          schema:
            $ref: "#/definitions/InvoiceResponse"
        401:
          description: "Not authorized"
        429:
          description: "Too many invoices requested. Max 3 invoices per Minute. Max 10 per Hour."
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
definitions:
  InvoiceResponse:
    type: "object"
    properties:
      payment_request:
        type: string
        description: Invoice Payment Request
...